refactor trust basis types and diff split
